Pickle the calculated database
