+++ .gitignore
@@ -1,2 +1,3 @@
 dist/
 frontend/.parcel-cache
+api/*.pickle
